Add batching option to statsd-client-cpp
This change is to improve the efficiency of communication between the client and
stated daemon.
The main thread will be just generating stats (in the format of character string)
in memory and a worker threads (a.k.a flusher) will be running in the background
to asynchronously sending batched stats to stated daemon via UDP package.
The flusher wake up every second and send all accumulated stats to the statsd
daemon.